Refactor to the new path structure
