+++ src/main/java/org/folio/rest/migration/BibMigration.java
@@ -40,8 +40,8 @@ import org.folio.rest.jaxrs.model.mod_source_record_storage.RecordModel;
 import org.folio.rest.migration.config.model.Database;
 import org.folio.rest.migration.mapping.InstanceMapper;
 import org.folio.rest.migration.model.BibRecord;
-import org.folio.rest.migration.model.request.BibContext;
-import org.folio.rest.migration.model.request.BibJob;
+import org.folio.rest.migration.model.request.bib.BibContext;
+import org.folio.rest.migration.model.request.bib.BibJob;
 import org.folio.rest.migration.service.MigrationService;
 import org.folio.rest.migration.utility.TimingUtility;
 import org.folio.rest.model.ReferenceLink;
