+++ src/main/java/org/folio/rest/migration/HoldingMigration.java
@@ -23,10 +23,10 @@ import org.folio.rest.jaxrs.model.Locations;
 import org.folio.rest.migration.config.model.Database;
 import org.folio.rest.migration.mapping.HoldingMapper;
 import org.folio.rest.migration.model.HoldingRecord;
-import org.folio.rest.migration.model.request.HoldingContext;
-import org.folio.rest.migration.model.request.HoldingDefaults;
-import org.folio.rest.migration.model.request.HoldingJob;
-import org.folio.rest.migration.model.request.HoldingMaps;
+import org.folio.rest.migration.model.request.holding.HoldingContext;
+import org.folio.rest.migration.model.request.holding.HoldingDefaults;
+import org.folio.rest.migration.model.request.holding.HoldingJob;
+import org.folio.rest.migration.model.request.holding.HoldingMaps;
 import org.folio.rest.migration.service.MigrationService;
 import org.folio.rest.migration.utility.TimingUtility;
 import org.folio.rest.model.ReferenceLink;
