+++ src/main/java/org/folio/rest/migration/ItemMigration.java
@@ -35,9 +35,9 @@ import org.folio.rest.jaxrs.model.Statisticalcodes;
 import org.folio.rest.jaxrs.model.CirculationNote.NoteType;
 import org.folio.rest.migration.config.model.Database;
 import org.folio.rest.migration.model.ItemRecord;
-import org.folio.rest.migration.model.request.ItemContext;
-import org.folio.rest.migration.model.request.ItemDefaults;
-import org.folio.rest.migration.model.request.ItemJob;
+import org.folio.rest.migration.model.request.item.ItemContext;
+import org.folio.rest.migration.model.request.item.ItemDefaults;
+import org.folio.rest.migration.model.request.item.ItemJob;
 import org.folio.rest.migration.service.MigrationService;
 import org.folio.rest.migration.utility.TimingUtility;
 import org.folio.rest.model.ReferenceLink;
