+++ src/main/java/org/folio/rest/migration/UserMigration.java
@@ -26,10 +26,10 @@ import org.folio.rest.jaxrs.model.Usergroups;
 import org.folio.rest.migration.config.model.Database;
 import org.folio.rest.migration.model.UserAddressRecord;
 import org.folio.rest.migration.model.UserRecord;
-import org.folio.rest.migration.model.request.UserContext;
-import org.folio.rest.migration.model.request.UserDefaults;
-import org.folio.rest.migration.model.request.UserJob;
-import org.folio.rest.migration.model.request.UserMaps;
+import org.folio.rest.migration.model.request.user.UserContext;
+import org.folio.rest.migration.model.request.user.UserDefaults;
+import org.folio.rest.migration.model.request.user.UserJob;
+import org.folio.rest.migration.model.request.user.UserMaps;
 import org.folio.rest.migration.service.MigrationService;
 import org.folio.rest.migration.utility.TimingUtility;
 import org.folio.rest.model.ReferenceLink;
