+++ src/main/java/org/folio/rest/migration/VendorMigration.java
@@ -28,10 +28,10 @@ import org.folio.rest.migration.model.VendorAddressRecord;
 import org.folio.rest.migration.model.VendorAliasRecord;
 import org.folio.rest.migration.model.VendorPhoneRecord;
 import org.folio.rest.migration.model.VendorRecord;
-import org.folio.rest.migration.model.request.VendorContext;
-import org.folio.rest.migration.model.request.VendorDefaults;
-import org.folio.rest.migration.model.request.VendorJob;
-import org.folio.rest.migration.model.request.VendorMaps;
+import org.folio.rest.migration.model.request.vendor.VendorContext;
+import org.folio.rest.migration.model.request.vendor.VendorDefaults;
+import org.folio.rest.migration.model.request.vendor.VendorJob;
+import org.folio.rest.migration.model.request.vendor.VendorMaps;
 import org.folio.rest.migration.service.MigrationService;
 import org.folio.rest.migration.utility.TimingUtility;
 import org.folio.rest.model.ReferenceLink;
