+++ src/main/java/org/folio/rest/migration/controller/MigrationController.java
@@ -9,14 +9,14 @@ import org.folio.rest.migration.UserReferenceLinkMigration;
 import org.folio.rest.migration.VendorMigration;
 import org.folio.rest.migration.VendorReferenceLinkMigration;
 import org.folio.rest.migration.aspect.annotation.CreateReferenceLinkTypes;
-import org.folio.rest.migration.model.request.BibContext;
-import org.folio.rest.migration.model.request.HoldingContext;
-import org.folio.rest.migration.model.request.InventoryReferenceLinkContext;
-import org.folio.rest.migration.model.request.ItemContext;
-import org.folio.rest.migration.model.request.UserContext;
-import org.folio.rest.migration.model.request.UserReferenceLinkContext;
-import org.folio.rest.migration.model.request.VendorContext;
-import org.folio.rest.migration.model.request.VendorReferenceLinkContext;
+import org.folio.rest.migration.model.request.bib.BibContext;
+import org.folio.rest.migration.model.request.holding.HoldingContext;
+import org.folio.rest.migration.model.request.inventory.InventoryReferenceLinkContext;
+import org.folio.rest.migration.model.request.item.ItemContext;
+import org.folio.rest.migration.model.request.user.UserContext;
+import org.folio.rest.migration.model.request.user.UserReferenceLinkContext;
+import org.folio.rest.migration.model.request.vendor.VendorContext;
+import org.folio.rest.migration.model.request.vendor.VendorReferenceLinkContext;
 import org.folio.rest.migration.service.MigrationService;
 import org.folio.spring.tenant.annotation.TenantHeader;
 import org.springframework.beans.factory.annotation.Autowired;
