+++ src/main/java/org/folio/rest/migration/model/VendorRecord.java
@@ -13,7 +13,7 @@ import org.folio.rest.jaxrs.model.acq_models.mod_orgs.schemas.Email;
 import org.folio.rest.jaxrs.model.acq_models.mod_orgs.schemas.Organization;
 import org.folio.rest.jaxrs.model.acq_models.mod_orgs.schemas.PhoneNumber;
 import org.folio.rest.jaxrs.model.acq_models.mod_orgs.schemas.Url;
-import org.folio.rest.migration.model.request.VendorDefaults;
+import org.folio.rest.migration.model.request.vendor.VendorDefaults;
 
 public class VendorRecord extends AbstractVendorRecord {
