+++ src/main/java/org/folio/rest/migration/model/request/bib/BibJob.java
@@ -1,4 +1,4 @@
-package org.folio.rest.migration.model.request;
+package org.folio.rest.migration.model.request.bib;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -6,6 +6,7 @@ import java.util.Map;
 import javax.validation.constraints.NotNull;
 
 import org.folio.rest.jaxrs.model.mod_data_import_converter_storage.JobProfile;
+import org.folio.rest.migration.model.request.AbstractJob;
 
 public class BibJob extends AbstractJob {
