+++ src/main/java/org/folio/rest/migration/model/request/item/ItemJob.java
@@ -1,8 +1,10 @@
-package org.folio.rest.migration.model.request;
+package org.folio.rest.migration.model.request.item;
 
 import java.util.Map;
 
 import javax.validation.constraints.NotNull;
+
+import org.folio.rest.migration.model.request.AbstractJob;
 
 public class ItemJob extends AbstractJob {
