+++ src/main/java/org/folio/rest/migration/model/request/user/UserDefaults.java
@@ -1,4 +1,4 @@
-package org.folio.rest.migration.model.request;
+package org.folio.rest.migration.model.request.user;
 
 public class UserDefaults {
