+++ src/main/java/org/folio/rest/migration/model/request/user/UserMaps.java
@@ -1,4 +1,4 @@
-package org.folio.rest.migration.model.request;
+package org.folio.rest.migration.model.request.user;
 
 import java.util.HashMap;
 import java.util.Map;
