+++ src/main/java/org/folio/rest/migration/model/request/user/UserReferenceLinkExtraction.java
@@ -1,4 +1,6 @@
-package org.folio.rest.migration.model.request;
+package org.folio.rest.migration.model.request.user;
+
+import org.folio.rest.migration.model.request.AbstractExtraction;
 
 public class UserReferenceLinkExtraction extends AbstractExtraction {
