+++ src/main/java/org/folio/rest/migration/model/request/user/UserReferenceLinkJob.java
@@ -1,7 +1,9 @@
-package org.folio.rest.migration.model.request;
+package org.folio.rest.migration.model.request.user;
 
 import java.util.HashMap;
 import java.util.Map;
+
+import org.folio.rest.migration.model.request.AbstractJob;
 
 public class UserReferenceLinkJob extends AbstractJob {
